Move more lexing logic out of the parser, expose a few functions at the module level
