+++ mochi/parser/__init__.py
@@ -1,0 +1,2 @@
+from mochi.parser.lexer import lex, REPL_CONTINUE
+from mochi.parser.parser import parse, Symbol, Keyword, get_temp_name
